be/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
joblib>=1.0.0
pandas>=1.3.0
scikit-learn>=0.24.2
python-multipart
requests
google-genai
dotenv
